requirements.txt
libusb1==3.0.0
loguru==0.7.0
ndeflib==0.3.3
nfcpy==1.0.4
prompt-toolkit==3.0.38
pyDes==2.0.1
pyserial==3.5
wcwidth==0.2.6
